adjust folders
