+++ src/pages/home/admin.js
@@ -1,9 +1,9 @@
 import React, {useContext} from "react";
-import AuthContext from "../context/AuthContext";
 import {Card, Container, Table} from "react-bootstrap";
+import AuthContext from "../../context/AuthContext";
 
-function Usuario() {
-    const {user} = useContext(AuthContext);
+function Admin() {
+    const { users } = useContext(AuthContext);
 
     function checkEstado(user) {
         if (user.estado === "Activo") {
@@ -14,18 +14,17 @@ function Usuario() {
     }
 
     return (
-        <Container className="d-flex flex-column justify-content-center align-items-center"
-                   style={{marginTop: "100px"}}>
+        <Container className="d-flex flex-column justify-content-center align-items-center" style={{ marginTop: "100px" }}>
             <Card className="p-4 shadow">
                 <Card.Img
                     variant="top"
                     src="/images/ash.png"
                     alt="Profile"
-                    style={{height: "400px", width: "300px", objectFit: "cover", borderRadius: "10px", margin: "auto"}}
+                    style={{ height: "400px", width: "300px", objectFit: "cover", borderRadius: "10px", margin: "auto" }}
                 />
 
                 <Card.Body>
-                    <Card.Title className="text-center">Información de Usuario</Card.Title>
+                    <Card.Title className="text-center">Información de Usuarios</Card.Title>
 
                     <Table striped bordered hover className="text-center">
                         <thead>
@@ -39,18 +38,20 @@ function Usuario() {
                         </tr>
                         </thead>
                         <tbody>
-                        <tr key={user.id}>
-                            <td>{user.id}</td>
-                            <td>{user.nombre}</td>
-                            <td>{user.email}</td>
-                            <td>{user.role}</td>
-                            <td>
-                                <div className="form-check form-switch">
-                                    {checkEstado(user)}
-                                </div>
-                            </td>
-                            <td>{user.fechaCreacion}</td>
-                        </tr>
+                        {users.map((user) => (
+                            <tr key={user.id}>
+                                <td>{user.id}</td>
+                                <td>{user.nombre}</td>
+                                <td>{user.email}</td>
+                                <td>{user.role}</td>
+                                <td>
+                                    <div className="form-check form-switch">
+                                        {checkEstado(user)}
+                                    </div>
+                                </td>
+                                <td>{user.fechaCreacion}</td>
+                            </tr>
+                        ))}
                         </tbody>
                     </Table>
                 </Card.Body>
@@ -59,4 +60,4 @@ function Usuario() {
     );
 }
 
-export default Usuario;
+export default Admin;
