+++ src/pages/home/usuario.js
@@ -1,9 +1,9 @@
 import React, {useContext} from "react";
+import AuthContext from "../../context/AuthContext";
 import {Card, Container, Table} from "react-bootstrap";
-import AuthContext from "../context/AuthContext";
 
-function Admin() {
-    const { users } = useContext(AuthContext);
+function Usuario() {
+    const {user} = useContext(AuthContext);
 
     function checkEstado(user) {
         if (user.estado === "Activo") {
@@ -14,17 +14,18 @@ function Admin() {
     }
 
     return (
-        <Container className="d-flex flex-column justify-content-center align-items-center" style={{ marginTop: "100px" }}>
+        <Container className="d-flex flex-column justify-content-center align-items-center"
+                   style={{marginTop: "100px"}}>
             <Card className="p-4 shadow">
                 <Card.Img
                     variant="top"
                     src="/images/ash.png"
                     alt="Profile"
-                    style={{ height: "400px", width: "300px", objectFit: "cover", borderRadius: "10px", margin: "auto" }}
+                    style={{height: "400px", width: "300px", objectFit: "cover", borderRadius: "10px", margin: "auto"}}
                 />
 
                 <Card.Body>
-                    <Card.Title className="text-center">Información de Usuarios</Card.Title>
+                    <Card.Title className="text-center">Información de Usuario</Card.Title>
 
                     <Table striped bordered hover className="text-center">
                         <thead>
@@ -38,20 +39,18 @@ function Admin() {
                         </tr>
                         </thead>
                         <tbody>
-                        {users.map((user) => (
-                            <tr key={user.id}>
-                                <td>{user.id}</td>
-                                <td>{user.nombre}</td>
-                                <td>{user.email}</td>
-                                <td>{user.role}</td>
-                                <td>
-                                    <div className="form-check form-switch">
-                                        {checkEstado(user)}
-                                    </div>
-                                </td>
-                                <td>{user.fechaCreacion}</td>
-                            </tr>
-                        ))}
+                        <tr key={user.id}>
+                            <td>{user.id}</td>
+                            <td>{user.nombre}</td>
+                            <td>{user.email}</td>
+                            <td>{user.role}</td>
+                            <td>
+                                <div className="form-check form-switch">
+                                    {checkEstado(user)}
+                                </div>
+                            </td>
+                            <td>{user.fechaCreacion}</td>
+                        </tr>
                         </tbody>
                     </Table>
                 </Card.Body>
@@ -60,4 +59,4 @@ function Admin() {
     );
 }
 
-export default Admin;
+export default Usuario;
